requirements.txt
mysql-connector-python==8.2.0
paho-mqtt==1.6.1
python-dotenv==1.0.0